# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "CreatePaymentRequest": {
        "description": "Request to create a payment via YooKassa.\n\npaymentMethod values: bank_card, sbp, sberbank, tinkoff_bank, yoo_money\nIf None, YooKassa will show all available payment methods.",
        "properties": {
          "paymentMethod": {
            "anyOf": [
              {
                "enum": [
                  "bank_card",
                  "sbp",
                  "sberbank",
                  "tinkoff_bank",
                  "yoo_money"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Paymentmethod"
          }
        },
        "title": "CreatePaymentRequest",
        "type": "object"
      },
...
      "FeedbackRequest": {
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "images": {
            "items": {
              "maxLength": 2000000,
              "type": "string"
            },
            "maxItems": 5,
            "title": "Images",
            "type": "array"
          },
          "text": {
            "maxLength": 10000,
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "email",
          "text"
        ],
        "title": "FeedbackRequest",
        "type": "object"
      },
...
      "GooglePurchase": {
        "properties": {
          "productId": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Productid",
            "type": "string"
          },
          "purchaseToken": {
            "maxLength": 4096,
            "minLength": 1,
            "title": "Purchasetoken",
            "type": "string"
          }
        },
        "required": [
          "purchaseToken",
          "productId"
        ],
        "title": "GooglePurchase",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "MessageResponse": {
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "title": "MessageResponse",
        "type": "object"
      },
...
      "PaymentConfirmationResponse": {
        "properties": {
          "confirmationUrl": {
            "title": "Confirmationurl",
            "type": "string"
          }
        },
        "required": [
          "confirmationUrl"
        ],
        "title": "PaymentConfirmationResponse",
        "type": "object"
      },
      "PremiumStatusResponse": {
        "properties": {
          "premium": {
            "title": "Premium",
            "type": "boolean"
          }
        },
        "required": [
          "premium"
        ],
        "title": "PremiumStatusResponse",
        "type": "object"
      },
...
      "StatusResponse": {
        "properties": {
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "StatusResponse",
        "type": "object"
      },
      "SubscriptionStatusResponse": {
        "properties": {
          "active": {
            "title": "Active",
            "type": "boolean"
          },
          "expiresAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expiresat"
          },
          "platform": {
            "anyOf": [
              {
                "enum": [
                  "yookassa",
                  "google_play",
                  "legacy"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Platform"
          }
        },
        "required": [
          "active",
          "platform",
          "expiresAt"
        ],
        "title": "SubscriptionStatusResponse",
        "type": "object"
      },
      "UserResponse": {
        "properties": {
          "email": {
            "title": "Email",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "premium": {
            "title": "Premium",
            "type": "boolean"
          },
          "subscriptionExpires": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subscriptionexpires"
          }
        },
        "required": [
          "id",
          "email",
          "premium"
        ],
        "title": "UserResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "YooKassaWebhookObject": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id"
        ],
        "title": "YooKassaWebhookObject",
        "type": "object"
      },
      "YooKassaWebhookRequest": {
        "properties": {
          "event": {
            "enum": [
              "payment.succeeded",
              "payment.canceled"
            ],
            "title": "Event",
            "type": "string"
          },
          "object": {
            "$ref": "#/components/schemas/YooKassaWebhookObject"
          },
          "type": {
            "const": "notification",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "event",
          "object"
        ],
        "title": "YooKassaWebhookRequest",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Food Spend Tracker API",
    "version": "1.5.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/subscription": {
      "get": {
        "description": "Get current subscription status.",
        "operationId": "get_subscription_api_subscription_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Subscription",
        "tags": [
          "Subscription"
        ]
      }
    },
    "/api/subscription/google/verify": {
      "post": {
        "operationId": "verify_google_purchase_api_subscription_google_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GooglePurchase"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Verify Google Purchase",
        "tags": [
          "Subscription"
        ]
      }
    },
    "/api/subscription/is_premium": {
      "get": {
        "description": "Check if user has active premium subscription.",
        "operationId": "is_premium_api_subscription_is_premium_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PremiumStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Is Premium",
        "tags": [
          "Subscription"
        ]
      }
    },
    "/api/subscription/payment": {
      "post": {
        "description": "Create YooKassa payment for premium subscription.",
        "operationId": "create_payment_api_subscription_payment_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/CreatePaymentRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentConfirmationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Payment",
        "tags": [
          "Subscription"
        ]
      }
    },
    "/api/subscription/yookassa/webhook": {
      "post": {
        "description": "Verify the current payment at YooKassa, then apply the event once.",
        "operationId": "webhook_api_subscription_yookassa_webhook_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/YooKassaWebhookRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Webhook",
        "tags": [
          "Subscription"
        ]
      }
    },
...
    "/api/users/me": {
      "get": {
        "operationId": "get_me_api_users_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Me",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/users/send-feedback": {
      "post": {
        "description": "Send feedback with optional images to the app owner",
        "operationId": "send_feedback_api_users_send_feedback_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Send Feedback",
        "tags": [
          "Users"
        ]
      }
    },
...
  }
}